requirements/optional.txt
mock==5.0.2
Markdown==3.4.3
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
XlsxWriter==3.1.2
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
amqp==5.1.1
celery==5.3.1
billiard==4.1.0


python-dateutil==2.8.2
cffi==1.15.1
cryptography==41.0.1
defusedxml==0.7.1
et-xmlfile==1.1.0
importlib-metadata==6.7.0
jwt==1.3.1
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
ruamel.yaml.clib==0.2.7
shortuuid==1.0.11
six==1.16.0
zipp==3.15.0
